Use float for all decimal types

+++ ziplime/assets/entities/futures_contract.py
@@ -1,6 +1,5 @@
 from dataclasses import dataclass
 from datetime import datetime
-from decimal import Decimal
 
 from ziplime.assets.entities.asset import Asset
 
@@ -12,8 +11,8 @@ class FuturesContract(Asset):
     symbol_mapping: dict[str, str]
     notice_date: datetime.date
     expiration_date: datetime.date
-    multiplier: Decimal
-    tick_size: Decimal
+    multiplier: float
+    tick_size: float
 
     def get_symbol_by_exchange(self, exchange_name: str) -> str | None:
         return self.symbol_mapping.get(exchange_name, None)
