+++ ziplime/domain/portfolio.py
@@ -1,6 +1,5 @@
 import datetime
 from dataclasses import dataclass, field
-from decimal import Decimal
 
 from ziplime.assets.entities.asset import Asset
 from ziplime.domain.position import Position
@@ -9,14 +8,14 @@ from ziplime.domain.position import Position
 @dataclass
 class Portfolio:
     # capital_used: float
-    cash_flow: Decimal
-    starting_cash: Decimal
-    portfolio_value: Decimal
-    pnl: Decimal
-    returns: Decimal
-    cash: Decimal
-    positions_value: Decimal
-    positions_exposure: Decimal
+    cash_flow: float
+    starting_cash: float
+    portfolio_value: float
+    pnl: float
+    returns: float
+    cash: float
+    positions_value: float
+    positions_exposure: float
     positions: dict[Asset, Position] = field(default_factory=dict)
     start_date: datetime.datetime | None = None
