+++ ziplime/finance/domain/position.py
@@ -17,7 +17,6 @@ Position Tracking
 """
 import dataclasses
 import datetime
-from decimal import Decimal
 
 from ziplime.assets.entities.asset import Asset
 
@@ -26,8 +25,8 @@ from ziplime.assets.entities.asset import Asset
 class Position:
     asset: Asset
     amount: int
-    cost_basis: Decimal
-    last_sale_price: Decimal
+    cost_basis: float
+    last_sale_price: float
     last_sale_date: datetime.datetime | None
 
     def __repr__(self):
