+++ ziplime/finance/domain/transaction.py
@@ -1,5 +1,4 @@
 import datetime
-from decimal import Decimal
 
 from ziplime.assets.entities.asset import Asset
 from ziplime.protocol import DataSourceType
@@ -7,9 +6,9 @@ from ziplime.protocol import DataSourceType
 
 class Transaction:
 
-    def __init__(self, asset: Asset, amount: int, dt: datetime.datetime, price: Decimal, order_id: str,
+    def __init__(self, asset: Asset, amount: int, dt: datetime.datetime, price: float, order_id: str,
                  exchange_name: str,
-                 commission: Decimal | None = None):
+                 commission: float | None = None):
         self.asset = asset
         self.amount = amount
         self.exchange_name=exchange_name
