+++ ziplime/trading/entities/orders/market_order_request.py
@@ -1,5 +1,4 @@
 import datetime
-from decimal import Decimal
 
 from ziplime.trading.entities.orders.order_request import OrderRequest
 from ziplime.trading.entities.trading_pair import TradingPair
@@ -9,7 +8,7 @@ from ziplime.trading.enums.order_type import OrderType
 
 class MarketOrderRequest(OrderRequest):
 
-    def __init__(self, order_id: str, trading_pair: TradingPair, order_side: OrderSide, quantity: Decimal,
+    def __init__(self, order_id: str, trading_pair: TradingPair, order_side: OrderSide, quantity: float,
                  exchange_name: str, creation_date: datetime.datetime
                  ):
         super().__init__(order_id=order_id, trading_pair=trading_pair, exchange_name=exchange_name,
