+++ ziplime/trading/entities/orders/order_request.py
@@ -1,6 +1,5 @@
 import datetime
 from dataclasses import dataclass
-from decimal import Decimal
 
 from ziplime.trading.entities.trading_pair import TradingPair
 from ziplime.trading.enums.order_type import OrderType
@@ -12,6 +11,6 @@ class OrderRequest:
     trading_pair: TradingPair
     order_type: OrderType
     creation_date: datetime.datetime
-    quantity:Decimal
+    quantity: float
 
     exchange_name:str
